Mountain Jumps
Given N mountains and the heights to each one, calculate the minimum energy needed to go from mountain 1 to mountain N. Going from mountain A to mountain B costs the difference in height between the tallest of the two and the shortest of the two. You can move to mountain A from mountain A - 1 or from mountain A - 2, provided that A - 1 and A -2 are > 0, respectively.