Longest Increasing Subsequence problem
Given an array with A(between 1 and 2500, inclusive) integers(each between 1 and 10000, inclusive), find the longest strictly increasing subsequence of the array. A subsequence is a set of elements that appear in the same relative order as they do in the array. It doesn't have to be composed of adjacent elements.