Shortest Routes - Multiple trips
Given a number N of cities, a number E of weighted roads between cities and asked a number Q of trips, inform for each trip Qi the minimum cost of the trip. If it is not possible to make the trip, print -1. I implemented this using a mix of Dijkstra and Floyd-Warshall, depending on how large the number of trips is compared to the number of cities